+++ src/main/java/ru/yandex/practicum/filmorate/controller/MpaController.java
@@ -25,10 +25,11 @@ public class MpaController {
     }
 
     @GetMapping("/mpa/{id}")
-    public ResponseEntity<MpaRating> getMpaById(@PathVariable Long id) {
+    public ResponseEntity<MpaRating> getMpaById(@PathVariable Integer id) { // Изменяем Long на Integer
         log.info("Received GET request for /mpa/{}", id);
+        // Аналогично GenreController, обработка NotFoundException может быть делегирована ErrorHandler.
         try {
-            MpaRating mpaRating = mpaRatingService.getMpaById(id);
+            MpaRating mpaRating = mpaRatingService.getMpaById(Long.valueOf(id));
             log.info("Returning MPA rating: {}", mpaRating);
             return ResponseEntity.ok(mpaRating);
         } catch (NotFoundException e) {
